28629 Progress - Changed image-viewer to file-viewer - Added support to view PDFs in file-viewer page
